requirements.txt
# Core dependencies for NL to SQL Agent - Phase 1
anthropic==0.39.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-community==0.3.5

# Utilities
tabulate==0.9.0
colorama==0.4.6
httpx==0.27.2